+++ src/main/php/ValueReader.php
@@ -493,52 +493,6 @@ class ValueReader implements valuereader\CommonValueReader
     }
 
     /**
-     * returns value if it denotes a path to an existing file, and null otherwise
-     *
-     * This should be used with greatest care in web environments as it only
-     * checks if the file exists, but not if there are any rights to access
-     * the specific file. It also does not prevent constructions which would
-     * allow an attacker to reach e.g. /etc/passwd via ../../ constructions.
-     *
-     * @api
-     * @param   string  $basePath  optional  base path where file must reside in
-     * @return  string
-     * @since   2.0.0
-     */
-    public function ifIsFile($basePath = null)
-    {
-        $path = ((null != $basePath) ? ($basePath . '/') : (''));
-        return $this->when(
-                new \stubbles\predicate\IsExistingFile($basePath),
-                'FILE_INVALID',
-                ['PATH' => $path . $this->param->value()]
-        );
-    }
-
-    /**
-     * returns value if it denotes a path to an existing directory, and null otherwise
-     *
-     * This should be used with greatest care in web environments as it only
-     * checks if the directory exists, but not if there are any rights to access
-     * the specific directory. It also does not prevent constructions which would
-     * allow an attacker to reach a certain directory via ../../ constructions.
-     *
-     * @api
-     * @param   string  $basePath  optional  base path where directory must reside in
-     * @return  string
-     * @since   2.0.0
-     */
-    public function ifIsDirectory($basePath = null)
-    {
-        $path = ((null != $basePath) ? ($basePath . '/') : (''));
-        return $this->when(
-                new \stubbles\predicate\IsExistingDirectory($basePath),
-                'DIRECTORY_INVALID',
-                ['PATH' => $path . $this->param->value()]
-        );
-    }
-
-    /**
      * returns param value when given predicate evaluates to true
      *
      * If value does not satisfy the predicate return value will be null.